+++ versions.tf
@@ -2,8 +2,8 @@ terraform {
   required_providers {
     cato = {
       source  = "catonetworks/cato"
-      version = ">= 0.0.24"
+      version = ">= 0.0.54"
     }
   }
-  required_version = ">= 1.4"
+  required_version = ">= 1.13"
 }