+++ CHANGELOG.md
@@ -2,3 +2,9 @@
 
 ### Features
 - Initial release of generic Cato IPSec module
+
+## 0.0.2 (2025-11-13)
+
+### Features
+- Updated site_location removingunnessary fields as required. 
+- Used updated native ipsec resource to define init_message and auth_message and removed nested raw API calls and need for auth credentials in variables.